:wrench: Fixed #1 Security warning

+++ src/main/java/me/drison64/mypin/Objects/Action/Action_Wait.java
@@ -63,7 +63,7 @@ public class Action_Wait extends Action {
                 runNext(data, line, block, player);
 
             }
-        }, delay * 20);
+        }, (long) delay * 20);
 
     }
